OpenVPN 1.0.2 build 98 (iOS)
Ported iOS client and OpenVPN 3 core to ARM-64.
Now building a "fat binary" with Xcode 5.0.1 that
targets arm7, arm7s, and arm64.

Outstanding issues:

* IPv6 doesn't route through tunnel on iOS7
* Client doesn't install on iOS 5.1.1.

+++ client/ovpncli.cpp
@@ -722,7 +722,7 @@ namespace openvpn {
 
     OPENVPN_CLIENT_EXPORT int OpenVPNClient::stats_n()
     {
-      return MySessionStats::combined_n();
+      return (int)MySessionStats::combined_n();
     }
 
     OPENVPN_CLIENT_EXPORT std::string OpenVPNClient::stats_name(int index)
@@ -799,7 +799,7 @@ namespace openvpn {
 	    if (lpr.defined())
 	      {
 		const Time::Duration dur = Time::now() - lpr;
-		const unsigned int delta = dur.to_binary_ms();
+		const unsigned int delta = (unsigned int)dur.to_binary_ms();
 		if (delta <= 60*60*24*1024) // only define for time periods <= 1 day
 		  ret.lastPacketReceived = delta;
 	      }
